+++ libs/dimcli/cli.cpp
@@ -1103,9 +1103,8 @@ vector<const char *> Cli::toPtrArgv(const vector<string> & args) {
 //   is preserved. All other chars are preserved.
 //
 // When escaping it's simplest to not quote and just escape the following:
-//   Must: | & ; < > ( ) $ ` \ " ' SP TAB LF
+//   Must: | & ; < > ( ) $ ` \ " ' SP TAB CR LF FF VTAB
 //   Should: * ? [ # ~ = %
-//
 //===========================================================================
 // static
 vector<string> Cli::toGlibArgv(const string & cmdline) {
@@ -1219,10 +1218,9 @@ IN_DQUOTE:
 //===========================================================================
 // Rules from libiberty's buildargv().
 //
-// Arguments split on unquoted whitespace (" \t\r\n\f\v")
+// Arguments are split on whitespace (" \t\r\n\f\v") unless quoted or escaped
 //  - backslashes: always escapes the following character.
 //  - single quotes and double quotes: escape each other and whitespace.
-//
 //===========================================================================
 // static
 vector<string> Cli::toGnuArgv(const string & cmdline) {
@@ -1307,7 +1305,14 @@ IN_QUOTED:
 //   - an odd number followed by a double quote adds one backslash for each
 //     pair, the last one is tossed, and the quote is added to the argument.
 //   - any number not followed by a double quote are literals.
-//
+//===========================================================================
+static void appendBackslashes(string & arg, int & backslashes) {
+    if (backslashes) {
+        arg.append(backslashes, '\\');
+        backslashes = 0;
+    }
+}
+
 //===========================================================================
 // static
 vector<string> Cli::toWindowsArgv(const string & cmdline) {
@@ -1317,13 +1322,6 @@ vector<string> Cli::toWindowsArgv(const string & cmdline) {
 
     string arg;
     int backslashes = 0;
-
-    auto appendBackslashes = [&arg, &backslashes]() {
-        if (backslashes) {
-            arg.append(backslashes, '\\');
-            backslashes = 0;
-        }
-    };
 
 IN_GAP:
     while (cur < last) {
@@ -1359,17 +1357,17 @@ IN_UNQUOTED:
         case '\t':
         case '\r':
         case '\n':
-            appendBackslashes();
+            appendBackslashes(arg, backslashes);
             out.push_back(move(arg));
             arg.clear();
             goto IN_GAP;
         default:
-            appendBackslashes();
+            appendBackslashes(arg, backslashes);
             arg += ch;
             break;
         }
     }
-    appendBackslashes();
+    appendBackslashes(arg, backslashes);
     out.push_back(move(arg));
     return out;
 
@@ -1389,14 +1387,141 @@ IN_QUOTED:
             }
             goto IN_UNQUOTED;
         default:
-            appendBackslashes();
+            appendBackslashes(arg, backslashes);
             arg += ch;
             break;
         }
     }
-    appendBackslashes();
+    appendBackslashes(arg, backslashes);
     out.push_back(move(arg));
     return out;
+}
+
+
+/****************************************************************************
+*
+*   Join argv to form a command line
+*
+***/
+
+//===========================================================================
+// static
+string Cli::toCmdline(const vector<string> & args) {
+    auto ptrs = toPtrArgv(args);
+    return toCmdline(ptrs.size(), ptrs.data());
+}
+
+//===========================================================================
+// static
+string Cli::toCmdline(size_t argc, char * argv[]) {
+#if defined(_WIN32)
+    return toWindowsCmdline(argc, argv);
+#else
+    return toGnuCmdline(argc, argv);
+#endif
+}
+
+//===========================================================================
+// static
+string Cli::toCmdline(size_t argc, const char * argv[]) {
+    return toCmdline(argc, (char **) argv);
+}
+
+//===========================================================================
+// static
+string Cli::toGlibCmdline(size_t, char * argv[]) {
+    string out;
+    if (!*argv)
+        return out;
+    for (;;) {
+        for (auto ptr = *argv; *ptr; ++ptr) {
+            switch (*ptr) {
+            // Must escape
+            case '|': case '&': case ';': case '<': case '>': case '(':
+            case ')': case '$': case '`': case '\\': case '"': case '\'':
+            case ' ': case '\t': case '\r': case '\n': case '\f': case '\v':
+            // Should escape
+            case '*': case '?': case '[': case '#': case '~': case '=':
+            case '%':
+                out += '\\';
+            }
+            out += *ptr;
+        }
+        if (!*++argv)
+            return out;
+        out += ' ';
+    }
+}
+
+//===========================================================================
+// static
+string Cli::toGnuCmdline(size_t, char * argv[]) {
+    string out;
+    if (!*argv)
+        return out;
+    for (;;) {
+        for (auto ptr = *argv; *ptr; ++ptr) {
+            switch (*ptr) {
+            case ' ': case '\t': case '\r': case '\n': case '\f': case '\v':
+            case '\\': case '\'': case '"':
+                out += '\\';
+            }
+            out += *ptr;
+        }
+        if (!*++argv)
+            return out;
+        out += ' ';
+    }
+}
+
+//===========================================================================
+// static
+string Cli::toWindowsCmdline(size_t, char * argv[]) {
+    string out;
+    if (!*argv)
+        return out;
+
+    for (;;) {
+        auto base = out.size();
+        int backslashes = 0;
+        auto ptr = *argv;
+        for (; *ptr; ++ptr) {
+            switch (*ptr) {
+            case '\\': backslashes += 1; break;
+            case ' ':
+            case '\t': goto QUOTE;
+            case '"':
+                out.append(backslashes + 1, '\\');
+                backslashes = 0;
+                break;
+            default: backslashes = 0; break;
+            }
+            out += *ptr;
+        }
+        goto NEXT;
+
+    QUOTE:
+        backslashes = 0;
+        out.insert(base, 1, '"');
+        out += *ptr++;
+        for (; *ptr; ++ptr) {
+            switch (*ptr) {
+            case '\\': backslashes += 1; break;
+            case '"':
+                out.append(backslashes + 1, '\\');
+                backslashes = 0;
+                break;
+            default: backslashes = 0; break;
+            }
+            out += *ptr;
+        }
+        out += '"';
+
+    NEXT:
+        if (!*++argv)
+            return out;
+        out += ' ';
+    }
 }